+++ packages/block-library/src/template-part/edit/selection-modal/index.js
@@ -6,18 +6,20 @@ import { kebabCase } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { useCallback } from '@wordpress/element';
+import { useCallback, useState } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { store as noticesStore } from '@wordpress/notices';
 import { useDispatch } from '@wordpress/data';
 import { serialize } from '@wordpress/blocks';
 import { store as coreStore } from '@wordpress/core-data';
+import { Button } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import TemplatePartList from './template-part-list';
 import PatternsList from './patterns-list';
+import TitleModal from './title-modal';
 
 export default function TemplatePartSelectionModal( {
 	setAttributes,
@@ -29,6 +31,7 @@ export default function TemplatePartSelectionModal( {
 } ) {
 	const { createSuccessNotice } = useDispatch( noticesStore );
 	const { saveEntityRecord } = useDispatch( coreStore );
+	const [ showTitleModal, setShowTitleModal ] = useState( false );
 
 	const onTemplatePartSelect = useCallback( ( templatePart ) => {
 		setAttributes( {
@@ -49,8 +52,8 @@ export default function TemplatePartSelectionModal( {
 		onClose();
 	}, [] );
 
-	const onBlockPatternSelect = async (
-		startingBlocks = [],
+	const createFromBlocks = async (
+		blocks = [],
 		title = __( 'Untitled Template Part' )
 	) => {
 		// If we have `area` set from block attributes, means an exposed
@@ -60,7 +63,7 @@ export default function TemplatePartSelectionModal( {
 		const record = {
 			title,
 			slug: kebabCase( title ),
-			content: serialize( startingBlocks ),
+			content: serialize( blocks ),
 			// `area` is filterable on the server and defaults to `UNCATEGORIZED`
 			// if provided value is not allowed.
 			area,
@@ -75,28 +78,45 @@ export default function TemplatePartSelectionModal( {
 			theme: templatePart.theme,
 			area: undefined,
 		} );
+		onClose();
 	};
 
 	return (
-		<div className="block-library-template-part__selection">
-			<div>
-				<h2>{ __( 'Pick from the existing template parts' ) }</h2>
-				<TemplatePartList
-					area={ area }
-					templatePartId={ templatePartId }
-					onSelect={ onTemplatePartSelect }
-				/>
+		<>
+			<div className="block-library-template-part__selection-content">
+				<div>
+					<h2>{ __( 'Pick from the existing template parts' ) }</h2>
+					<TemplatePartList
+						area={ area }
+						templatePartId={ templatePartId }
+						onSelect={ onTemplatePartSelect }
+					/>
+				</div>
+
+				<div>
+					<h2>{ __( 'Pick from the existing patterns' ) }</h2>
+					<PatternsList
+						area={ area }
+						areaLabel={ areaLabel }
+						onSelect={ createFromBlocks }
+						clientId={ clientId }
+					/>
+				</div>
 			</div>
 
-			<div>
-				<h2>{ __( 'Pick from the existing patterns' ) }</h2>
-				<PatternsList
-					area={ area }
-					areaLabel={ areaLabel }
-					onSelect={ onBlockPatternSelect }
-					clientId={ clientId }
-				/>
+			<div className="block-library-template-part__selection-footer">
+				<Button isSecondary onClick={ () => setShowTitleModal( true ) }>
+					{ __( 'Start blank' ) }
+				</Button>
 			</div>
-		</div>
+
+			{ showTitleModal && (
+				<TitleModal
+					areaLabel={ areaLabel }
+					onClose={ () => setShowTitleModal( false ) }
+					onSubmit={ ( title ) => createFromBlocks( [], title ) }
+				/>
+			) }
+		</>
 	);
 }
