+++ packages/block-library/src/template-part/edit/selection-modal/patterns-list.js
@@ -6,16 +6,13 @@ import {
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
 import { useState, useMemo } from '@wordpress/element';
-import { __, sprintf } from '@wordpress/i18n';
-import {
-	TextControl,
-	Flex,
-	FlexItem,
-	Button,
-	Modal,
-} from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import { useAsyncList } from '@wordpress/compose';
+
+/**
+ * Internal dependencies
+ */
+import TitleModal from './title-modal';
 
 export default function PatternsList( {
 	area,
@@ -47,13 +44,11 @@ export default function PatternsList( {
 
 	// Restructure onCreate to set the blocks on local state.
 	// Add modal to confirm title and trigger onCreate.
-	const [ title, setTitle ] = useState( __( 'Untitled Template Part' ) );
 	const [ startingBlocks, setStartingBlocks ] = useState( [] );
 	const [ isTitleStep, setIsTitleStep ] = useState( false );
 	const shownPatterns = useAsyncList( filteredBlockPatterns );
 
-	const submitForCreation = ( event ) => {
-		event.preventDefault();
+	const onSubmitTitle = ( title ) => {
 		onSelect( startingBlocks, title );
 	};
 
@@ -69,39 +64,11 @@ export default function PatternsList( {
 			/>
 
 			{ isTitleStep && (
-				<Modal
-					title={ sprintf(
-						// Translators: %s as template part area title ("Header", "Footer", etc.).
-						__( 'Name and create your new %s' ),
-						areaLabel.toLowerCase()
-					) }
-					closeLabel={ __( 'Cancel' ) }
-					overlayClassName="wp-block-template-part__placeholder-create-new__title-form"
-					onRequestClose={ () => setIsTitleStep( false ) }
-				>
-					<form onSubmit={ submitForCreation }>
-						<TextControl
-							label={ __( 'Name' ) }
-							value={ title }
-							onChange={ setTitle }
-						/>
-						<Flex
-							className="wp-block-template-part__placeholder-create-new__title-form-actions"
-							justify="flex-end"
-						>
-							<FlexItem>
-								<Button
-									variant="primary"
-									type="submit"
-									disabled={ ! title.length }
-									aria-disabled={ ! title.length }
-								>
-									{ __( 'Create' ) }
-								</Button>
-							</FlexItem>
-						</Flex>
-					</form>
-				</Modal>
+				<TitleModal
+					areaLabel={ areaLabel }
+					onClose={ () => setIsTitleStep( false ) }
+					onSubmit={ onSubmitTitle }
+				/>
 			) }
 		</>
 	);
